+++ server/database/entities.go
@@ -22,9 +22,9 @@ type Player struct {
 
 type Throw struct {
 	Model
-	Points        uint8 `json:"points"`
-	Multiplicator uint8 `json:"multiplicator"`
-	PlayerId      uint  `json:"playerId"`
+	Points        uint16 `json:"points"`
+	Multiplicator uint16 `json:"multiplicator"`
+	PlayerId      uint   `json:"playerId"`
 }
 
 type Game struct {
